requirements.txt
# fastapi
# uvicorn
# pydantic
# python-dotenv
# requests
# pymupdf
# python-docx
# langchain
# langchain-community
# faiss-cpu
# email-validator
# openai


# scikit-learn


fastapi
uvicorn
pydantic
python-dotenv
numpy
requests
httpx
PyMuPDF
python-docx
pymupdf
email-validator
